% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuRosim2.R
\docType{package}
\name{neuRosim2}
\alias{neuRosim2}
\alias{neuRosim2-package}
\title{Functions to Generate fMRI Data Including Activated Data, Noise Data and Resting State Data}
\description{
The package allows users to generate fMRI time series or 4D data. Some high-level functions
are created for fast data generation with only a few arguments and a diversity of functions
to define activation and noise. For more advanced users it is possible to use the low-level
functions and manipulate the arguments.
}
\references{
Welvaert, M., Durnez, J., Moerkerke, B., Verdoolaege, G. and Rosseel, Y. (2011). \pkg{neuRosim}: An {R} Package for Generating fMRI Data. Journal of Statistical Software, 44(10), 1--18
}
\author{
\emph{Maintainer:} Monica Thieu \email{monica.thieu@columbia.edu}

Original authors:
\itemize{
  \item{Marijke Welvaert}
  \item{Joke Durnez}
  \item{Beatrijs Moerkerke}
  \item{Yves Rosseel}
  \item{Geert Verdoolaege}
}
}
