% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifydesign.R
\name{specifydesign}
\alias{specifydesign}
\title{Generate design matrix.}
\usage{
specifydesign(stim.function, TR, effectsize, conv = c("none", "gamma",
  "double-gamma", "Balloon"), param = NULL)
}
\arguments{
\item{stim.function}{A tibble output by \code{\link{stimfunction}} specifying onsets.}

\item{TR}{Repetition time in seconds.}

\item{effectsize}{List or number representing the effectsize in each condition.}

\item{conv}{Should the design matrix be convoluted, default is none.}

\item{param}{Parameters of the haemodynamic response function.
See \code{\link{gammaHRF}} and \code{\link{canonicalHRF}} for more details.}
}
\value{
A tibble with columns specifying the design.
}
\description{
Generates a design matrix to be used as a model for the simulated activation.
}
\examples{
os <- list(c(20,60),c(15,35))
d <- list(20, 10)
total <- 80
TR <- 2
out <- specifydesign(stimfunction(total, os, d, acc = 0.1), TR,
  effectsize=list(2,5), conv="double-gamma")
\dontshow{rm(os,d,total,TR,out)}

}
\seealso{
\code{\link{specifyregion}},\code{\link{gammaHRF}},\code{\link{canonicalHRF}},\code{\link{balloon}}
}
\keyword{low-level}
