% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{temporalnoise}
\alias{temporalnoise}
\title{Generate temporally correlated noise}
\usage{
temporalnoise(dim, nscan, sigma, rho = 0.2, template, verbose = TRUE)
}
\arguments{
\item{dim}{A vector specifying the dimensions of a 2D or 3D array.}

\item{nscan}{The number of scans in the dataset.}

\item{sigma}{The standard deviation of the noise.}

\item{rho}{The autocorrelation coefficients. The length of the vector determines the order of the autoregressive model.}

\item{template}{An array representing the anatomical structure or mask with dimensions equal to dim.}

\item{verbose}{Logical indicating if warnings should be printed.}
}
\value{
An array containing the noise with dimensions specified in \code{dim}.
}
\description{
Generates an autoregressive noise dataset with specified dimensions and standard deviation.
}
\examples{
d <- c(10,10,10)
sigma <- 5
nscan <- 100
rho <- c(0.3,-0.7)
out <- temporalnoise(d, nscan, sigma, rho, verbose=FALSE)
\dontshow{rm(d,sigma,nscan,rho,out)}

}
\seealso{
\code{\link{systemnoise}}, \code{\link{lowfreqdrift}}, \code{\link{physnoise}},
\code{\link{tasknoise}}, \code{\link{spatialnoise}}
}
\keyword{low-level}
\keyword{noise}
