% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gammaHRF}
\alias{gammaHRF}
\title{Single Gamma Haemodynamic response function}
\usage{
gammaHRF(x, FWHM = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Time vector in seconds.}

\item{FWHM}{Full Width Half Maximum of the Gamma variate function.}

\item{verbose}{If \code{TRUE}, warnings are displayed.}
}
\value{
Vector representing the values of the function for the given time vector and FWHM.
}
\description{
Specifies a Gamma variate haemodynamic response function for the given time vector and FWHM.
}
\examples{
t <- 1:100
out <- gammaHRF(t, verbose=FALSE)
\dontshow{rm(t,out)}

}
\references{
Buxton, RB, Uludag, K, Dubowitz, DJ and Liu, TT (2004). Modeling the hemodynamic response to brain activation. NeuroImage, 23, S220-S233.
}
\seealso{
\code{\link{canonicalHRF}}, \code{\link{balloon}}
}
\keyword{activation}
\keyword{low-level}
